We can use size_t here
This matches the memory allocation functions in libc.

+++ source/ac/memory.c
@@ -11,7 +11,7 @@
 #include "logging.h"
 
 static void*
-defaultAC_memmanage(void* ctxptr, void* old, uint32_t size)
+defaultAC_memmanage(void* ctxptr, void* old, size_t size)
 {
     (void)ctxptr;
     if (size > 0) {
